v1.1 update

+++ src/main/java/me/sleepyfish/rat/modules/impl/Fullbright.java
@@ -5,11 +5,11 @@ import me.sleepyfish.rat.modules.Module;
 /**
  * This class is from Rat Client.
  * WARNING: Unauthorized reproduction, skidding, or decompilation of this code is strictly prohibited.
- * @author Nexuscript 2024
+ * @author SleepyFish 2024
  */
 public class Fullbright extends Module {
 
-    private float oldGammaSetting;
+    private short oldGammaSetting;
 
     public Fullbright() {
         super("Fullbright", "Makes everything bright");
@@ -17,14 +17,14 @@ public class Fullbright extends Module {
 
     @Override
     public void onEnableEvent() {
-        this.oldGammaSetting = mc.gameSettings.gammaSetting;
+        this.oldGammaSetting = (short) mc.gameSettings.gammaSetting;
         mc.gameSettings.gammaSetting = 100F;
     }
 
     @Override
     public void onDisableEvent() {
         mc.gameSettings.gammaSetting = this.oldGammaSetting;
-        this.oldGammaSetting = 0F;
+        this.oldGammaSetting = 0;
     }
 
 }
